Bug-Fix: Waiting for the transaction status fixed
